Fix for issue 43:
The oracle codec test was incorrectly testing the oracle codec like it
was the mysql standard codec. It now correctly tests the oracle codec.

Additionally, this also breaks out the different tests into separate
tests. This is in accordance with my understanding of unit test best
practice of just testing one thing at a time. It also allows other
codecs to be tested even when earlier ones fail (the unix and windows
codecs were never tested because the oracle ones failed). Additionally
this makes it clear that just the oracle codec is failing and not the
entire codec subsystem.