Big nasty patch to back port the XMLEntityCodec to 1.4. This includes most of the functionality needed for the HTMLCodec fix which is next. This includes codec.HashTrieTest, util.NullSafe and util.CollectionsUtil.
Two new classes have been added:

codec.AbstractCodec:

This is a base abstract codec.Codec implementation to
ease porting. In 1.4 Codec is a interface and in 2.0 it is a abstract
class. Ports back to 1.4 ca use AbstractCodec as their base instead.

util.PrimWrap:

This is a simple class to wrap primitives in their java.lang classes. This
is here to help back porting of auto boxing code from the 2.0 branch. By
using this instead of new Character(), etc we can easily implement our
own fly weight caching of these objects as 1.5 does in it's auto boxing
if and when the overhead incurred in creating new objects each time
becomes a issue.