Synced HTML entity, HTML attribute, and JavaScript encoding with 2.0 branch.
The behavior between 1.4 and 2.0 was different before, and 2.0 stopped some attacks
that 1.4 did not.

Also, wrote it so that encodeCharacter returns void and takes a StringBuffer
so that it does not have to allocate so many strings. In my testing, this reduced
average run time for a test case that ran encodeForHTML, encodeForHTMLAttribute, and
then encodeForJavaScript to 1220ms from 3400ms.